Revert "* upgrade to React 17.x & function-based components"
This reverts commit 350a4f2f22092f64ad9ad80c3e1840a41fed1c9e.
