--- .babelrc
@@ -1,16 +1,0 @@
-{
-	"presets": [
-	  ["@babel/preset-env", {"modules": "auto"}],
-	  ["@babel/preset-react"]
-	],
-	"plugins": [
-	  "@babel/plugin-proposal-class-properties",
-	  "@babel/plugin-proposal-private-methods"
-	],
-    "env": {
-		"debug": {
-		  "sourceMaps": "inline",
-		  "retainLines": true
-		}
-	}
-}
