--- .dockerignore
@@ -1,7 +1,0 @@
-.git/
-.env*
-docker-compose.yml
-Dockerfile
-Makefile
-LICENSE
-README.md
