--- .env.example
@@ -1,3 +1,0 @@
-COMPOSE_PROJECT_NAME=smartlife-tui
-IMAGE=shellcatt/${COMPOSE_PROJECT_NAME}
-VERSION=latest
