+++ .gitignore
@@ -1,3 +1,1 @@
-.env
-node_modules/
-.trash/
+node_modules
