--- .vscode/launch.json
@@ -1,29 +1,0 @@
-{
-    "configurations": [
-    {
-        "type": "node",
-        "request": "attach",
-        "name": "Debug Babel-Node in Docker",
-        "port": 9229,
-        "localRoot": "${workspaceFolder}",
-        "remoteRoot": "/src",
-        "protocol": "inspector",
-        "runtimeArgs": [
-          "--nolazy",
-          "--debug-brk"
-        ],
-        "dockerServerReadyAction": {
-          "pattern": "Debugger listening on "
-        },
-        "env": {
-          "BABEL_ENV": "debug"
-        },
-        "outFiles": [
-          "${workspaceFolder}/dist/**/*.js"
-        ],
-        "skipFiles": [
-          "<node_internals>/**"
-        ],
-      }
-    ]
-}
