--- .vscode/tasks.json
@@ -1,15 +1,0 @@
-{
-	"version": "0.2.0",
-    "tasks": [
-		{
-			"label": "Docker: Run debugger",
-			"type": "shell",
-			"command": "docker compose up app-debug --build",
-			"problemMatcher": [],
-			"group": {
-				"kind": "build",
-				"isDefault": true
-			}
-		}
-	],
-}
