+++ README.md
@@ -3,17 +3,10 @@
 ## TODO
 
 - [x] create single-screen TUI (`react-blessed-contrib`)
-- [*] create separate components based on device types
-  - [x] Socket
-  - [x] Light
-  - [ ] Power Strip
-  - [ ] ...
-- [ ] implement local state (for caching & grouping)
-- [ ] implement authentication TUI flow (dialog) 
-- [ ] create connector multiple backends  based on availability
-  - [ ] Local state
-  - [ ] SmartLife API (`tuya-smartlife-api-node`)
-  - [ ] Tuya OpenAPI (`tuya-openapi-node`)
+- [ ] create separate components based on device types
+- [ ] connect to SmartLife API wrapper (`tuya-smartlife-api-node`)
+- [ ] implement authentication TUI (dialog) 
+- [ ] create local state (for grouping)
 - [ ] dockerize
 
 
@@ -22,3 +15,4 @@
 
 - https://smartathome.co.uk/smartlife/
 - https://www.alanwood.net/demos/webdings.html
+- https://www.alanwood.net/demos/wingdings.html
