--- docker-compose.yml
@@ -1,13 +1,0 @@
-services:
-  app:
-    build: .
-    volumes:
-      - .:/src
-
-  app-debug:
-    build: .
-    ports:
-      - "9229:9229"  # Debugging port
-    volumes:
-      - .:/src
-    command: ["npm", "run", "test:debug"]
