+++ package.json
@@ -1,13 +1,12 @@
 {
   "name": "smartlife-tui",
-  "version": "0.8.0",
+  "version": "0.7.0",
   "description": "Tuya SmartLife TUI",
+  "main": "tet.js",
   "scripts": {
-    "start": "babel-node index.js",
-    "start:mon": "babel-node index.js",
-    "test": "babel-node tests/index.js",
-    "test:debug": "BABEL_ENV=debug babel-node --inspect-brk=0.0.0.0:9229 tests/index.js",
-    "lint": "eslint ."
+    "start": "node index",
+    "lint": "eslint .",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
@@ -28,25 +27,18 @@
   },
   "homepage": "https://github.com/shellcatt/smartlife-tui#readme",
   "dependencies": {
-    "@babel/cli": "^7.26.4",
     "@babel/core": "^7.23.7",
-    "@babel/node": "^7.26.0",
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-proposal-private-methods": "^7.18.6",
-    "@babel/plugin-transform-modules-commonjs": "^7.26.3",
     "@babel/plugin-transform-runtime": "^7.23.7",
     "@babel/preset-env": "^7.23.8",
     "@babel/preset-react": "^7.23.3",
     "@babel/register": "^7.23.7",
+    "@use-it/interval": "^1.0.0",
     "blessed": "^0.1.81",
-    "blessed-contrib": "^4.11.0",
-    "blessed-tab-container": "^1.0.0",
     "chalk": "^5.3.0",
-    "colors": "^1.4.0",
     "lodash": "^4.17.21",
-    "lodash.upperfirst": "^4.3.1",
-    "react": "^17.0.2",
-    "react-blessed": "^0.7.2",
+    "react": "^16.14.0",
+    "react-blessed": "^0.2.1",
+    "react-blessed-contrib": "^0.2.1",
     "stop-runaway-react-effects": "^2.0.0"
   },
   "devDependencies": {
