--- src/utils/actions.js
@@ -1,1 +1,0 @@
-export const actionTitle = (s) => '⏻ ' + (s ? 'On' : 'Off');
