--- tests/index.js
@@ -1,5 +1,0 @@
-console.log('hola mundo');
-debugger;
-console.log('hola amigos');
-debugger;
-console.log('adios mundo');
